+++ packages/tinacms/src/utils/index.ts
@@ -19,8 +19,8 @@ export interface CreateClientProps {
   clientId?: string
   isLocalClient?: boolean
   tinaioConfig?: TinaIOConfig
-  owner?: string,
-  repo?: string,
+  owner?: string
+  repo?: string
   branch?: string
 }
 export const createClient = ({