+++ scripts/transformDeps.sh
@@ -1,0 +1,5 @@
+git checkout -b examples
+yarn swapDeps
+git commit -am "update deps"
+git push origin examples
+git checkout -